requirements-app.txt
aiohttp==3.13.1
beautifulsoup4==4.13.5
langchain==0.3.27
langchain-community==0.3.27
langgraph==0.6.4
langgraph-prebuilt==0.6.4
mcp==1.18.0
oci==2.161.1
pydantic==2.12.3
python-dotenv==1.1.1
requests==2.32.5
slack_bolt==1.23.0
slack_sdk==3.36.0
typing_extensions==4.15.0
urllib3==2.5.0
websocket-client==1.8.0
websockets==15.0.1
faiss-cpu==1.8.0
numpy==1.26.4
rank-bm25==0.2.2
sentence-transformers==3.0.1